+++ drift_dev/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 2.12.1-dev
 
 - Fix invalid types listed in `views` crashing the generator.
+- Fix indices in drift files crashing the generator when they index expressions.
 
 ## 2.12.0